feat(lock): add recipient by input
